requirements.txt
PyQt5>=5.15
pandas>=1.3.0
pyyaml>=5.4